requirements.txt
discord.py>=2.0.0
aiohttp>=3.8.0
lxml>=4.9.0
pyyaml>=6.0
python-dotenv>=1.0.0
beautifulsoup4>=4.11.0
aiodns>=3.0.0
cchardet>=2.1.7
uvloop>=0.17.0; platform_system != "Windows"